+++ src/aspect/fpsmonitor.h
@@ -8,19 +8,19 @@
 class FpsMonitor : public Qt3DCore::QComponent {
     Q_OBJECT
     Q_PROPERTY( int rollingMeanFrameCount READ rollingMeanFrameCount WRITE setRollingMeanFrameCount NOTIFY rollingMeanFrameCountChanged )
-    Q_PROPERTY( float framesPerSecond READ framesPerSecond NOTIFY framesPerSecondChanged )
+    Q_PROPERTY( double framesPerSecond READ framesPerSecond NOTIFY framesPerSecondChanged )
 
   public:
     explicit FpsMonitor( Qt3DCore::QNode* parent = nullptr );
 
-    float framesPerSecond() const;
+    double framesPerSecond() const;
     int rollingMeanFrameCount() const;
 
   public slots:
     void setRollingMeanFrameCount( int rollingMeanFrameCount );
 
   signals:
-    void framesPerSecondChanged( float framesPerSecond );
+    void framesPerSecondChanged( double framesPerSecond );
     void rollingMeanFrameCountChanged( int rollingMeanFrameCount );
 
   protected:
